pyproject.toml
[tool.poetry]
name = "sdjquizz"
version = "0.1.0"
description = ""
authors = ["Steve De Jongh <sdejongh@he2b.be>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
